submarine: Added tutorial and basic environment
The resources/ directory and the logo (submarine.svg) are directly
taken from the gsoc_submarine branch

The basic environment (submarine, whale, crown, background, the upper
and the lower gates) are added to the activity

Tutorials are used whenever the user arrives at any on the first three
levels. Each of those three activity contains the tutorials for the
engine, ballast tanks and rudders respectively. The tutorials are
implemented via the IntroMessage component which can be found under
core/IntroMessage.qml

Box2D elements are added to the submarine, upper and lower gates,
crown and the whale. The World will not be running unless the
tutorials are over.

Signed-off-by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>