submarine: Fix element position on screen change
Since the positions of the dynamic objects are being maintained via js,
the properties of the objects won't get affected on screen size change.
So, Activity.resizeElements() was called which will do the initial
setup again whenever an onWidthChange or onHeightChanged signal is
emitted

Two key points to be noted:
* resizeElements() should not be called the first time the level is
loaded (and hence the onWidthChanged and onHeightChanged signal is
emitted)
* Few elements (like the crown) will change it's visibility during the
progress of the levels, so we should not change that property while
we are correcting the properties of the object on screen change

The crown is added in the level, which will trigger the gate to open
(not implemented yet). When a level has no crown, the gate will open
on start of the level

The number of items in the dataset has been reduced to consider only
those which are needed to be updated after every level

Signed-off-by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>