submarine: submarine Box2D target moved to Image
The Box2D component of the submarine is now targetted to the Image of
the submarine, so that the box2d component can follow the player.
The rock's collider was given an width and height, also allowed it to
collide with the player submarine.

Changed the graphics of the whale when it is hit by the submarine, also
added the option to reset the properties of the whale.

Added reset option for the crown, to avoid being disappeared once the
player loses after collecting the crown

Signed-off-by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>