submarine: variable onYChanged animation
Previously the speed duration for the animation when the Y value of the
submarine was changed was fixed. This lead to a weird behaviour in
which the submarine would move up/down sharply when the diving planes
turned ineffective. This can be replicate by:
1. Start the activity
2. Statr filling up the ballast tanks
3. increase horizontal velocity
4. increase diving planes angles for some time
5. decrease diving planes angles to 0

The duration of the animation is now dependent on the amount of
distance the submarine needs to cover up on the Y axis, multiplied by a
terminalVelocityIndex

However, we do not need it on level resets, so in that case, the value
resetVerticalSpeed (currently set to 500) is used

Signed-off-by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>