submarine: inertia: bar: crown fixes
Added inertia to the horizontal movement of the submarine. The
submarine's horizontal velocity is always equal to submarine.velocity,
so whenever the value of the velocity is increased or decreased, instead
of directly changing the values, the value is linearly interpolated to
give an effect of inertia

The bar is now set to be always visible, since it is not longer hiding
any component of the control panel anymore

Fixed undefined behaviour arrising due to resetUpperGate() being called
on activity start even before the "items" is fully loaded. In that
scenario, the call is ignored

Signed-off-by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>