submarine: inc water fill rate, smoothed velocity
The water level in the Ballast tanks now gets filled up almost
instantly, since the tanks of a submarine is either empty of full. The
velocity of the submarine underwater (when no Diving planes are being
used) also decreases with increase in depth of the submarine (since
Buoyant force increases and (mg - Fb) starts to decrease.

So, vertical velocity is given by:

maxVelocity * ((amount of depth to be overcomed) / total depth)

Signed-off-by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>