family: Implemented Grid layout
Instead of using the Grid element of Qt, we divide the screen into
few horizontal and vertical parts. Vertically, we divide it into
three parts:
gen_1: for Generation 1
gen_2: for Generation 2
gen_3: for Generation 3

(Let's denote it by "v")

Horizontally, it is divided into few parts. Primary divisions are:
left, center and right.

The left and right is also divided into two parts each:
left_1, left_2, right_1, right_2

(Let's denote these by "h")

Depending on the need of a particular generation

For placing the nodes, we simply use (h,v) with the appropriate values.
For placing the edges, we use the nodeWidth and nodeHeight variables to
accurately place the end points on the circumference of the nodes.

It has been used in levels 1, 4 and 5

Signed-off-by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>