family: Use active&activeTo states for expert mode
Previously, the states of the nodes were always "deactive" for the
expert mode. Two variables "firstNodePointer" and "secondNodePointer"
is maintained to keep track of the first and the second node that is
been chosen as a pair by the user and is marked as "active" and
"activeTo" whenever the buttons are pressed.

When the answer is found out to be wrong, the two nodes are deactivated.
It is not performed on win since all the nodes are automatically
deactivated on start of a new level

Signed-off-by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>