family: Randomise levels for family_find_relatives
The levels for the family_find_relatives activity is shuffled to
increase the randomness of the levels in the activity. Initially, the
shuffledLevelIndex[] contains values from 0 to numberOfLevels - 1, which
is later shuffled using the shuffle() function. Everytime the activity
starts, the levels are arranged in a new order

Since the levels are shuffled, the question label cannot be obtained
using dataset.levelElements[bar.level - 1].answer[0] anymore, which is
now obtained via the js

Signed-off-by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>