digital_electricity: Add ok button to check result
Added an "ok" button to verify the answer instead of checking the answer
everytime the bulb glows up.

determiningComponentsIndex[] stores the indices of the components which
will determine the answer, which upto level 4 is just the index of the
digital light. In "checkAnswer()" function, the value of the determining
component is checked and the result is determined accordingly.

Signed-off-by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>