digital_electricity: level 12: NOT gate using NAND
Level 12 is aimed at teaching the user to create one gate using other
gates. In this level, the user is asked to create a NOT gate using a
NAND gate only.

For dfbce1eb "digital_electricity: Update answer checking" fixed problem
of result not showing up and switches not setting up to correct values
on wrong answer. This is fixed by added:

  updateComponent(switch3.index)
  items.bonus.bad('tux')

for each of the above levels.

Signed-off-by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>