digital_electricity: NOR gates, Comparator
Levels 16 - 19 aims at teaching the implementation of other gates using
NOR gate. They are explained as follows:
* Level 16: Implementing NOT gate using NOR gate
* Level 17: Implementing AND gate using NOR gate
* Level 18: Implementing OR gate using NOR gate
* Level 19: Implementing NAND gate using NOR gate

Level 20 introduces comparator, with a simple task requiring the use of
comparator and asks the user to compute the following result using it:

  S = A <= B

where A and B are two respective switchKeys and S is the output of the
digitalLight.

Signed-off-by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>