digital_electricity: Free mode: Add component from dataset
The components to be loaded on the availablePieces for the free mode was
being done by appending each of the componenets individually, which is
now changed to a much more generalised manner, by listing all the
components in the Dataset under the componentList variable.

It is loaded in the js by running through the array one at a time and
appending them in the availablePieces:

        items.availablePieces.model.append( {
            "imgName": componentList[i].imageName,
            "componentSrc": componentList[i].componentSource,
            "imgWidth": sizeMultiplier * componentList[i].width,
            "imgHeight": sizeMultiplier * componentList[i].height,
            "toolTipText": componentList[i].toolTipText
        })

level 2: it's -> its
level 22: made the intromessage more descriptive by detailing on the bcd
counter and it's output.

Signed-off-by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>