digital_electricity: Move in the playArea
Implemented moving in the playArea using the arrow keys. On pressing the
arrow keys, we check on which side we should be moving and if they are
restricted to the given frame, the corresponding move is made.

The positioning of the components should also be relative to the current
zoom ratio, which is implemented by:

  components[i].posX *= zoomRatio
  components[i].posY *= zoomRatio

whenever there is zoom in or out.

Signed-off-by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>