digital_electricity: Reduced size of dragged component
Previously, the size of the dragged component was constant regardless
of the current zoom amount. The size is determined under the toFull()
function in the DragListItem.qml, in which we have the width and the
height set as:

  width = fullWidth
  height = fullHeight

which is changed to:

   width = fullWidth * Activity.currentZoom
   height = fullHeight * Activity.currentZoom

to make the size of the dragged component dependent on the size of the
zoom amount.

Signed-off-by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>