digital_electricity: Allow hide/unhide tool icons
Previously, the tools icons were placed in a Grid with two rows and two
columns, which were too small for devices with small screen size. This
is replaced by a single "Tools" button, which shows/hides a menu
containing all the tools options on alternate clicking. This avoids the
problem of having the icons of small size, along with keeping a balance
between amount of free space available and number of clicks required to
access a tool (once or more than once).

The Tools are placed under a Flow, whose parent is a Rectangle which
serves as the container of the Tools. When the tools option is clicked
alternatively, the visibility of the container is changed accordingly.

Signed-off-by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>