digital_electricity: Add zoom icons in toolbar
Added zoom in and out icons in the Toolbar along with their active and
inactive states, depending on whether the zooming is possible or not.

For active state, the button's opacity is kept 1.0, and for inactive
state, the opacity is reduced to 0.5.

Initially, the zoom is set to it's default value in the init() method in
the js, and the "Zoom in" button is deactivated, while the zoom out
button is activated.

In the zoomIn() function, after calculating the final zoom amount, we
are checking whether the current zoom is the maximum allowed zoom or
not, and if so, we are setting the state of the zoom in button as
deactive.

Similarly in the zoomOut() function, after calculating the final zoom
amount, we are checking if the current zoom amount is same as that of
the minimum zoom amount, and if so, we are deactivating the zoom out
button.

Signed-off-by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>