digital_electricity: Used Text.Fit to fit text with button
In a612172 "Adapt tools text size with button", we updated the scale of
the text element depending on whether it exceeded the width of the
button or not. This has been changed to using Text.Fit to make sure that
the text remains within it's prescribed width, which is set to be the
same as that of the width of the button. Also, we align the text
horizontally using Text.AlignHCenter

Signed-off-by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>