digital_electricity: Correctly hide intromessage
Previously, the IntroMessage's visibility was forcefully turned off or
on depending on the availability of an intro message / current mode of
the activity. This is changed to the following:

As per 810e45e3, "core, IntroMessage: minor improvements to doxygen
docu", when the IntroMessage is needed to be hidden, it's index is set
to -1

  items.tutorialInstruction.index = -1

thus removing any kind of conflicts that may appear due to different
procedures of hiding the element.

Signed-off-by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>