Railroad, removed extra margins in drag & drog area
Extra margin was creating unexpected results on drag & drop, espacially
in top part of sample wagon row. Now to move a wagon from sample area
to answer row, it needs to be dropped completely in answer row. There is
a margin for dropping a wagon from answer row to handle unexpected drop of
wagon while replacing position of wagons in answer row.