background-music, fix music looping order
this fixes problem with music restarting at track 2 instead of 1 after the end of last track.

However this brings back the issue that when starting the app with music disabled,
re-enabling it only works after restarting the app.

Best would be to find a way to fix both,
but if not possible the remaining "restart-needed" issue seems less annoying
than the "skipped track" issue fixed in this case.